Throw if the websocket cant connect
